Extract clock module
