+++ lib/asset-loader/index.js
