+++ lib/pathfinder.js
@@ -1,6 +1,6 @@
 const _ = require('lodash')
 const { AStarFinder, Grid } = require('pathfinding')
-const { getAllEntities } = require('../game')
+const { getAllEntities } = require('./game')
 
 const finder = new AStarFinder()
